+++ spec/directions/move_custom_section_to_eoc_container/v1_spec.rb
@@ -144,6 +144,97 @@ RSpec.describe Kitchen::Directions::MoveCustomSectionToEocContainer do
         HTML
       )
     end
+
+    it 'adds a section-level wrapper' do
+      described_class.v1(
+        chapter: book_with_section_to_move.chapters.first,
+        metadata_source: metadata_element,
+        container_key: 'some-eoc-section',
+        uuid_key: '.some-eoc-section',
+        section_selector: 'section.some-eoc-section',
+        append_to: append_to,
+        wrap_section: true
+      )
+      expect(append_to.first('$.os-some-eoc-section-container')).to match_normalized_html(
+        <<~HTML
+          <div class="os-eoc os-some-eoc-section-container" data-type="composite-page" data-uuid-key=".some-eoc-section">
+            <h3 data-type="title">
+              <span class="os-text">Some Eoc Section</span>
+            </h3>
+            <div data-type="metadata" style="display: none;">
+              <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <div class="authors" id="authors_copy_1">Authors</div>
+              <div class="publishers" id="publishers_copy_1">Publishers</div>
+              <div class="print-style" id="print-style_copy_1">Print Style</div>
+              <div class="permissions" id="permissions_copy_1">Permissions</div>
+              <div data-type="subject" id="subject_copy_1">Subject</div>
+            </div>
+            <div class="os-section-area">
+              <section class="some-eoc-section" id="sectionId1">
+                <p>content</p>
+              </section>
+            </div>
+            <div class="os-section-area">
+              <section class="some-eoc-section" id="sectionId2">
+                <p>content</p>
+              </section>
+            </div>
+            <div class="os-section-area">
+              <section class="some-eoc-section" id="sectionId3">
+                <p>content</p>
+              </section>
+            </div>
+          </div>
+        HTML
+      )
+    end
+
+    it 'adds a content-level wrapper' do
+      described_class.v1(
+        chapter: book_with_section_to_move.chapters.first,
+        metadata_source: metadata_element,
+        container_key: 'some-eoc-section',
+        uuid_key: '.some-eoc-section',
+        section_selector: 'section.some-eoc-section',
+        append_to: append_to,
+        wrap_section: true,
+        wrap_content: true
+      )
+      expect(append_to.first('$.os-some-eoc-section-container')).to match_normalized_html(
+        <<~HTML
+          <div class="os-eoc os-some-eoc-section-container" data-type="composite-page" data-uuid-key=".some-eoc-section">
+            <h3 data-type="title">
+              <span class="os-text">Some Eoc Section</span>
+            </h3>
+            <div data-type="metadata" style="display: none;">
+              <h1 data-type="document-title" itemprop="name">Some Eoc Section</h1>
+              <div class="authors" id="authors_copy_1">Authors</div>
+              <div class="publishers" id="publishers_copy_1">Publishers</div>
+              <div class="print-style" id="print-style_copy_1">Print Style</div>
+              <div class="permissions" id="permissions_copy_1">Permissions</div>
+              <div data-type="subject" id="subject_copy_1">Subject</div>
+            </div>
+            <div class="os-some-eoc-section">
+              <div class="os-section-area">
+                <section class="some-eoc-section" id="sectionId1">
+                  <p>content</p>
+                </section>
+              </div>
+              <div class="os-section-area">
+                <section class="some-eoc-section" id="sectionId2">
+                  <p>content</p>
+                </section>
+              </div>
+              <div class="os-section-area">
+                <section class="some-eoc-section" id="sectionId3">
+                  <p>content</p>
+                </section>
+              </div>
+            </div>
+          </div>
+        HTML
+      )
+    end
   end
 
   context 'when append_to is nil' do