+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Refactor `MoveExercisesToEOC::V2` & `BakeChapterKeyConcepts` to use `MoveCustomSectionToEocContainer` (minor)
+* Add wrapper support for `MoveCustomSectionToEocContainer` (minor)
 * Create `BakeHandbook` direction (minor)
 * Add `ExampleElement#titles_to_rename` & refactor `BakeExample` (patch)
 * Create `BakeUnitPageTitle` (minor)